requirements.txt
# Core data processing
numpy>=1.26.0
pandas>=2.0.0
pyarrow>=14.0.0
openpyxl>=3.1.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# HTTP / API clients
requests>=2.31.0
httpx>=0.27.0
kaggle>=1.6.0
kagglehub>=0.2.0

# Data quality & validation
rapidfuzz>=3.6.0
unicodedata2>=15.1.0
pydantic>=2.5.0
pandera>=0.18.0
tenacity>=8.2.0

# Logging & utilities
loguru>=0.7.0
tqdm>=4.66.0
python-dotenv>=1.0.0
